Move profile image to media image package
